Simplify number:buy and remove numbers:buy
This change makes it easier to buy a number quickly
and also simplifies the underlying numberBuy code.

Usage is now:

number:buy [number_pattern]

If number_pattern is a full number that number will
be purchased.

If it’s not a full number and a --country_id option
is set then a search is performed using number_pattern
as the pattern.

number:buy 5 --country_id GB

Or, just by the first number for a country:

number:buy --country_id US

numbers:buy (plural) has been removed. It threw
an exception anyway.