Remove vxml cli commands
With the shutdown of the call API, the link:vxml and unlink:vxml
commands can be removed as they are no longer used.

Note that this information might be referenced in other areas, ie. the
verbose numbers list screen, but this is OK as users may want to see
that a number is still linked to vxml in order to identify changes.